# apps/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "IoT 全栈开发工作室的招新管理后台 API。\n统一响应体是 {code, message, data}；除公开端点（注册/登录/发码/刷新/登出）外都需要 Authorization: Bearer \u003caccess_token\u003e。",
        "title": "iot-pillot API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/admin/admins": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "列出当前具备管理权限的账号。名单真源在 Redis。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理端"
                ],
                "summary": "管理员名单",
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.AdminListResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    },
                    "403": {
                        "description": "非管理员",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "把已注册用户加入管理员名单，并撤销其现有会话（旧令牌的 role 已过时，需重新登录）。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理端"
                ],
                "summary": "提升为管理员",
                "parameters": [
                    {
                        "description": "目标用户",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AdminUserIDReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.AdminMutationResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误或用户不存在",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    },
                    "403": {
                        "description": "非管理员",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/admins/{user_id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "把用户移出管理员名单，并撤销其现有会话。服务端强制 403 读的就是这份名单，因此立即生效。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理端"
                ],
                "summary": "撤销管理员",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "目标用户 ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.AdminMutationResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    },
                    "403": {
                        "description": "非管理员",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/mail-templates": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "列出全部邮件模板；每个模板附带它用到的变量名，便于管理端提示。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理端-邮件"
                ],
                "summary": "邮件模板列表",
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.MailModelListResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    },
                    "403": {
                        "description": "非管理员",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "主题与正文都支持 {{变量}} 占位符；括号不成对会被拒绝（那是会静默发出错信的错误）。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理端-邮件"
                ],
                "summary": "新建邮件模板",
                "parameters": [
                    {
                        "description": "模板内容",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.MailTemplateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.MailModelResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误或模板名/类型重复",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/mail-templates/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理端-邮件"
                ],
                "summary": "更新邮件模板",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "模板 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "模板内容",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.MailTemplateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.MailModelResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误、模板不存在或名称重复",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "只删模板，不动已发出的邮件记录——记录是审计凭据，不该因为模板被删就消失。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理端-邮件"
                ],
                "summary": "删除邮件模板",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "模板 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    },
                    "400": {
                        "description": "模板不存在",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/mails": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "按发送时间倒序分页。可选 keyword（主题/收件人模糊匹配）与 from/to（日期，含端点）过滤；total 为过滤后的总数。只返回记录元信息，不返回正文。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理端-邮件"
                ],
                "summary": "发信记录",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码，从 1 开始",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页条数，最大 100",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "模糊搜索主题与收件人邮箱，最长 100",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "起始日期 YYYY-MM-DD（含当天）",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束日期 YYYY-MM-DD（含当天）",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.MailRecordListResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "日期格式错误",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/mails/batch-delete": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "管理员清理记录用；返回实际删除的行数，ids 中已不存在的 id 会被跳过。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理端-邮件"
                ],
                "summary": "批量删除发信记录",
                "parameters": [
                    {
                        "description": "待删除的记录 id 列表",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.MailBatchDeleteReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.MailBatchDeleteResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/mails/send": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "变量不传也行：服务端会用收件人资料（姓名/邮箱/班级/方向/学号/轮次）预填，请求里的同键覆盖。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理端-邮件"
                ],
                "summary": "按 userID 发送邮件",
                "parameters": [
                    {
                        "description": "发送参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.SendMailToUserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.MailSendResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误、模板不存在、缺少变量或用户没有邮箱",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/mails/send-bulk": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "逐条独立投递，单个失败不影响其他收件人；响应里分别给出成功与失败清单。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理端-邮件"
                ],
                "summary": "群发邮件（同一模板，逐个替换内容）",
                "parameters": [
                    {
                        "description": "群发参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.SendMailBulkReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功（含逐条结果）",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.MailBulkSendResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误、模板不存在或收件人超过上限",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/mails/send-by-email": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "用于发给还没注册的报名者；邮箱能对上已注册用户时会按用户记录。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理端-邮件"
                ],
                "summary": "按邮箱地址发送邮件",
                "parameters": [
                    {
                        "description": "发送参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.SendMailToEmailReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.MailSendResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误、模板不存在或缺少变量",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "分页读取已注册用户，并标注其当前是否为管理员。响应不含密码哈希。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理端"
                ],
                "summary": "用户列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码，从 1 开始",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页条数，最大 100",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.AdminUserListResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    },
                    "403": {
                        "description": "非管理员",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    }
                }
            }
        },
...
        "/api/v1/logout": {
            "post": {
                "description": "退出登录接口（需要 access token 和 refresh token）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证模块"
                ],
                "summary": "用户退出登录接口",
                "parameters": [
                    {
                        "description": "退出登录请求参数（refresh_token）",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.RefreshReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "退出登录成功",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Result"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "request.AdminUserIDReq": {
            "type": "object",
            "required": [
                "user_id"
            ],
            "properties": {
                "user_id": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
...
        "response.AdminListResp": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.AdminUserResp"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "response.AdminMutationResp": {
            "type": "object",
            "properties": {
                "is_admin": {
                    "type": "boolean"
                },
                "session_revoked": {
                    "type": "boolean"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "response.Result": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务码：0 成功，非 0 失败",
                    "type": "integer"
                },
                "data": {
                    "description": "业务数据"
                },
                "message": {
                    "description": "提示信息",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "登录得到的 access_token。Swagger UI 的 Authorize 里只填 token 本身，前缀由 UI 补。",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}